mengganti spatie/laravel-pdf dgn barryvdh/laravel-dompdf - part 8

+++ app/Http/Controllers/BarangController.php
@@ -440,7 +440,9 @@ class BarangController extends Controller implements HasMiddleware
                 ->setOptions(['enable_php' => true]);
 
             $output = $pdf->output();
-            Storage::disk('pdfs')->delete($namafile);
+            if (Storage::disk('pdfs')->exists($namafile)) {
+                Storage::disk('pdfs')->delete($namafile);
+            }
             Storage::disk('pdfs')->put($namafile, $output);
 
             return response()->json([
