+++ app/Http/Controllers/SaleInvoiceController.php
@@ -158,7 +158,9 @@ class SaleInvoiceController extends Controller implements HasMiddleware
                 ->setOptions(['enable_php' => true]);
 
             $output = $pdf->output();
-            Storage::disk('pdfs')->delete($namafile);
+            if (Storage::disk('pdfs')->exists($namafile)) {
+                Storage::disk('pdfs')->delete($namafile);
+            }
             Storage::disk('pdfs')->put($namafile, $output);
 
             return response()->json([
@@ -194,7 +196,9 @@ class SaleInvoiceController extends Controller implements HasMiddleware
                 ->setOptions(['enable_php' => true]);
 
             $output = $pdf->output();
-            Storage::disk('pdfs')->delete($namafile);
+            if (Storage::disk('pdfs')->exists($namafile)) {
+                Storage::disk('pdfs')->delete($namafile);
+            }
             Storage::disk('pdfs')->put($namafile, $output);
 
             return response()->json([
