+++ app/Http/Controllers/StockAdjustmentController.php
@@ -277,7 +277,9 @@ class StockAdjustmentController extends Controller implements HasMiddleware
                 ->setOptions(['enable_php' => true]);
 
             $output = $pdf->output();
-            Storage::disk('pdfs')->delete($namafile);
+            if (Storage::disk('pdfs')->exists($namafile)) {
+                Storage::disk('pdfs')->delete($namafile);
+            }
             Storage::disk('pdfs')->put($namafile, $output);
 
             return response()->json([
